HS-95-05: dictation through the desk — the cockpit is a window, the voice is real
DictationCore carries the whole daily cockpit (9 sections) as a
host-agnostic core; the flat route is a thin wrapper; the desk hosts it in
a window via SurfaceWindows. Every dictation exit rewires to the shell:
the Dictate start chip, the DeskChrome room, the Pullout's 'Dictate about
this' (which scopes the window — hosts resolve the subject to a product
label, never a raw ref), and the inspector's project-context. The app
shell delegates navigate into desk/shell.ts (setShellNavigator /
openSurfaceOr) so cores and chrome fall back to a legacy route without
router imports; with 'dictate' registered the fallback is dead weight
until HS-95-08 deletes it. The Try-it utterance gains its standing-rule
speak-to-fill mic; mic authority is single-owner (a second capture takes
the mic and the first's tracks stop — pinned by micOwner.test.ts).

Proof (evidence-story-05.md + assets/dictation-voice-1440.png): suite
258/258; the dictation walk on the production bundle — the chip and the
Pullout open the window IN PLACE (URL stays on the desk), the scope chip
reads 'About Untitled meeting', the flat route still answers, and the
REAL voice path lands: a say-generated wav through Chromium's fake mic
device, transcribed by the hub's real Whisper into the utterance box.

PMO-Story: HS-95-05
PMO-Contract-Digest: sha256:c9ffe15d1a986bfb8d5c2b2ab79cefe9dcb6506b73a21f2ff7f979746d146f1b
