+++ web/src/App.tsx
@@ -1,12 +1,17 @@
-import { Suspense } from "react";
-import { Navigate, Route, Routes, useLocation } from "react-router-dom";
+import { Suspense, useEffect } from "react";
+import { Navigate, Route, Routes, useLocation, useNavigate } from "react-router-dom";
 import { AppShell } from "./components/AppShell";
 import { ErrorBoundary } from "./components/ErrorBoundary";
 import { Skeleton } from "./components/signal/Signal";
 import { PRODUCT_ROUTES } from "./routes";
+import { setShellNavigator } from "./desk/shell";
 
 export function App() {
   const location = useLocation();
+  const navigate = useNavigate();
+  useEffect(() => {
+    setShellNavigator((href) => navigate(href));
+  }, [navigate]);
   const route = PRODUCT_ROUTES.find((item) => item.path === location.pathname);
   const immersive = route?.immersive ?? false;
   return (
